cifs: deprecate 'enable_negotiate_signing' module param
We can blindly send a negotiate signing context on every negotiate call,
as it doesn't affect the current behaviour; even though we set AES-GMAC
as the first signing algorithm (i.e. our preferred algorithm), it will
use AES-CMAC if:

a) the server doesn't respond to signing negotiations
b) the server responds that it only supports AES-GMAC

So this is a safe change.

Throw a warning if 'enable_negotiate_signing=0' was explicitly set, but
ignore it and set it back to true.

Signed-off-by: Enzo Matsumiya <ematsumiya@suse.de>
Signed-off-by: Steve French <stfrench@microsoft.com>